+++ test/functional-tests/node-transformers/rename-properties-transformers/rename-properties-transformer/RenamePropertiesTransformer.spec.ts
@@ -294,7 +294,12 @@ describe('RenamePropertiesTransformer', () => {
             });
 
             describe('Variant #8: integration with `splitStrings` option', () => {
-                const propertyRegExp: RegExp = /'a': *'long' *\+ *'Prop' *\+ *'erty' *\+ *'Valu' *\+ *'e'/;
+                const propertyRegExp: RegExp = new RegExp(
+                    'const foo *= *{' +
+                        '\'a\': *\'long\' *\\+ *\'Prop\' *\\+ *\'erty\' *\\+ *\'Valu\' *\\+ *\'e\'' +
+                    '};' +
+                    'foo\\[\'a\'];'
+                );
 
                 let obfuscatedCode: string;
 
@@ -315,6 +320,31 @@ describe('RenamePropertiesTransformer', () => {
 
                 it('Should rename property before `splitStrings` option will applied', () => {
                     assert.match(obfuscatedCode, propertyRegExp);
+                });
+            });
+        });
+
+        describe('Ignored literal node type', () => {
+            describe('Variant #1: boolean literal node', () => {
+                const regExp: RegExp = /var obj *= *{}; *obj\[!!\[]] *= *0x1;/;
+
+
+                let obfuscatedCode: string;
+
+                before(() => {
+                    const code: string = readFileAsString(__dirname + '/fixtures/boolean-literal-node.js');
+
+                    obfuscatedCode = JavaScriptObfuscator.obfuscate(
+                        code,
+                        {
+                            ...NO_ADDITIONAL_NODES_PRESET,
+                            renameProperties: true
+                        }
+                    ).getObfuscatedCode();
+                });
+
+                it('Match #1: should skip literal property with invalid type', () => {
+                    assert.match(obfuscatedCode, regExp);
                 });
             });
         });